+++ yierdis-protocol-codec/src/main/java/yier/bubu/redis/protocol/json/JsonBoolean.java
@@ -1,0 +1,4 @@
+package yier.bubu.redis.protocol.json;
+
+public record JsonBoolean(boolean value) implements JsonValue {
+}