+++ yierdis-protocol-codec/src/main/java/yier/bubu/redis/protocol/json/JsonLong.java
@@ -1,0 +1,4 @@
+package yier.bubu.redis.protocol.json;
+
+public record JsonLong(long value) implements JsonNumber {
+}