+++ yierdis-protocol-codec/src/main/java/yier/bubu/redis/protocol/json/JsonNumber.java
@@ -1,0 +1,4 @@
+package yier.bubu.redis.protocol.json;
+
+public sealed interface JsonNumber extends JsonValue permits JsonLong, JsonDouble {
+}