+++ yierdis-protocol/src/main/java/yier/bubu/redis/protocol/json/JsonBoolean.java
@@ -1,5 +1,0 @@
-package yier.bubu.redis.protocol.json;
-
-public record JsonBoolean(boolean value) implements JsonValue {
-}
-