+++ yierdis-protocol/src/main/java/yier/bubu/redis/protocol/json/JsonLong.java
@@ -1,5 +1,0 @@
-package yier.bubu.redis.protocol.json;
-
-public record JsonLong(long value) implements JsonNumber {
-}
-