+++ yierdis-protocol/src/main/java/yier/bubu/redis/protocol/json/JsonNumber.java
@@ -1,5 +1,0 @@
-package yier.bubu.redis.protocol.json;
-
-public sealed interface JsonNumber extends JsonValue permits JsonLong, JsonDouble {
-}
-