+++ yierdis-protocol/src/main/java/yier/bubu/redis/protocol/reply/ReplyBoolean.java
@@ -1,8 +1,0 @@
-package yier.bubu.redis.protocol.reply;
-
-/**
- * IR boolean 值。
- */
-public record ReplyBoolean(boolean value) implements ReplyValue {
-}
-